Fixes using AppDelegateAdapter instead of AppDelegate directly
Fixes using AppDelegateAdapter instead of AppDelegate directly
