+++ Branch Monster Factor/Branch_Monster_FactorApp.swift
@@ -8,8 +8,11 @@
 import SwiftUI
 import SwiftData
 
-
+@main
 struct Branch_Monster_FactorApp: App {
+    
+    @UIApplicationDelegateAdaptor(AppDelegateAdapter.self) var appDelegate
+    
     var sharedModelContainer: ModelContainer = {
         let schema = Schema([
             Item.self,
@@ -25,7 +28,6 @@ struct Branch_Monster_FactorApp: App {
 
     var body: some Scene {
         WindowGroup {
-            //ContentView()
             OnboardingScreen()
         }
         .modelContainer(sharedModelContainer)
